Implement generator
Implement a generator that adds clues to the puzzle if it is not
solvable.

Enhance the solver so that it can use clue-wide clues to solve,
instead of only hints specific to a single position.

Define a criteria for a good puzzle and generate puzzles until
one meets the criteria.